Create a Quarto html report, Bioinfo-group styled

Create a Quarto html report, Bioinfo-group styled

Usage

create_bioinfo_html(
  file_name = NULL,
  report_folder = NULL,
  ext_name = "bioinfo-html",
  expose_aux_files = TRUE
)

Arguments

file_name

Character string, the name of the file for the report - no need to specify the qmd extension.

report_folder

Character string, pointing to the folder where the report and its associated files should be created. Defaults to NULL, which falls back to the current working directory. A sensible value for this variable could the "vignettes", if used in the context of a package in development

ext_name

Character string, defaulting to "bioinfo-html". Other values (possibly in other functions) would be "bioinfo-revealjs" or similar...

expose_aux_files

Logical value, whether to expose&copy the auxiliary files (scss and bib) into the toplevel of the specified folder.

Value

Invisibly, the path to the created quarto source file for the report

See also

Examples

temp_folder <- tempdir()
create_bioinfo_html(file_name = "myreport",
                    report_folder = temp_folder)
#> ℹ Using report folder, /var/folders/l7/zzmm12b945z_f71ljwfqj7z00000gn/T//RtmpxTdGib
#> ℹ Created '_extensions' folder into /var/folders/l7/zzmm12b945z_f71ljwfqj7z00000gn/T//RtmpxTdGib
#> ℹ bioinfo-html was installed to _extensions folder in the specified report directory.
#> ℹ Copied auxiliary files into the toplevel report folder.
#> You might want to edit the corresponding lines in the yaml section
#> ☐ Edit
#>   /var/folders/l7/zzmm12b945z_f71ljwfqj7z00000gn/T/RtmpxTdGib/myreport.qmd.
#> ✔ Have fun with your report - available inside the folder /var/folders/l7/zzmm12b945z_f71ljwfqj7z00000gn/T//RtmpxTdGib